requirements.txt
# Core dependencies
pandas>=2.0.0
chardet>=5.0.0
fastmcp>=1.0.0
psutil>=5.9.0
sqlalchemy>=2.0.0
pytest>=7.0.0
crewai
crewai[tools]
crewai-tools[mcp]
